Earthquake List
A simple Android application that will read JSON data from an
earthquakes API and populate a RecyclerView within the main activity.
Main ask was to refrain from using 3rd party libraries and code to
handle these requests and responses.

If the magnitude of the quake is equal or greater than 8, then
distinguish these visually from other earthquakes in the list.

On click of the recyclerView item, open a Map fragment where I get the
latitude and longitude from the JSON and show on the map.

Also, includes Unit tests and UI tests